Optimized code and added fine shuffle mode as the extra shuffle was too aggressive for small review counts.
