+++ src/hoochie_mama/const.py
@@ -4,5 +4,6 @@
 # License: GNU GPL, version 3 or later; http://www.gnu.org/copyleft/gpl.html
 
 
-from . import prefmenu
-from . import hoochieMama
+from anki import version
+ANKI21 = version.startswith("2.1.")
+CCBC = version.endswith("ccbc")
